add loop for PSO

+++ decToBin.m
@@ -1,9 +1,9 @@
 function bin = decToBin(dec)
-    temp = dec;
+    temp = int64(dec);
     flag = 0;
     j = 1;
     while flag ~= 1
-        bin(j) = mod(temp, 2);
+        bin(j) = int64(mod(temp, 2));
         temp = (temp - bin(j))/2;
         j = j + 1;
         if temp == 1
